Allow customize http vhost config using HttpdCustomization.CustomServiceConfigSecret
This change allows to customize the httpd vhost config using this parameter
to specify a secret that contains service config data. The content of each
provided snippet gets rendered as a go template and placed into
/etc/httpd/conf/httpd_custom_<endpoint>_<key> .
At the end of the vhost config in the default httpd template these custom
configs get included using `Include conf/httpd_custom_<endpoint>_*`.

For information on how sections in httpd configuration get merged, check section
"How the sections are merged" in https://httpd.apache.org/docs/current/sections.html#merging

All possible parameters which can be use in a template can be looked up
in the <service>-config-data secret of the service like:
$ oc get secret -n openstack keystone-config-data -o json | jq -r .data.templatingParameters | base64 -d

or in the running pod of the service in the file:
$ cat /var/lib/config-data/default/templatingParameters

The content is a versioned dump of the parameters of the service operator, like:

~~~
v1:
    DatabaseConnection: mysql+pymysql://user:pwd@openstack.openstack.svc/keystone?read_default_file=/etc/my.cnf
    KeystoneEndpointInternal: https://keystone-internal.openstack.svc:5000
    KeystoneEndpointPublic: https://keystone-public-openstack.apps-crc.testing
    ProcessNumber: 3
    TransportURL: rabbit://user:pwd@rabbitmq.openstack.svc:5671/?ssl=1
    VHosts:
        internal:
            Override: false
            SSLCertificateFile: /etc/pki/tls/certs/internal.crt
            SSLCertificateKeyFile: /etc/pki/tls/private/internal.key
            ServerName: keystone-internal.openstack.svc
            TLS: true
        public:
            Override: false
            SSLCertificateFile: /etc/pki/tls/certs/public.crt
            SSLCertificateKeyFile: /etc/pki/tls/private/public.key
            ServerName: keystone-public.openstack.svc
            TLS: true
    enableSecureRBAC: true
    fernetMaxActiveKeys: 5
    memcachedServers: memcached-0.memcached.openstack.svc:11212
    memcachedServersWithInet: inet:[memcached-0.memcached.openstack.svc]:11211
    memcachedTLS: true
~~~

Depends-On: https://github.com/openstack-k8s-operators/lib-common/pull/591

Signed-off-by: Martin Schuppert <mschuppert@redhat.com>